+++ ring/ring_poly.go
@@ -3,12 +3,12 @@ package ring
 import (
 	"encoding/binary"
 	"errors"
-	"github.com/tuneinsight/lattigo/v3/utils"
 )
 
 // Poly is the structure that contains the coefficients of a polynomial.
 type Poly struct {
-	Coeffs  [][]uint64 // Coefficients in CRT representation
+	Coeffs  [][]uint64 // Dimension-2 slice of coefficients (re-slice of Buff)
+	Buff    []uint64   // Dimension-1 slice of coefficient
 	IsNTT   bool
 	IsMForm bool
 }
@@ -16,10 +16,13 @@ type Poly struct {
 // NewPoly creates a new polynomial with N coefficients set to zero and Level+1 moduli.
 func NewPoly(N, Level int) (pol *Poly) {
 	pol = new(Poly)
+
+	pol.Buff = make([]uint64, N*(Level+1))
 	pol.Coeffs = make([][]uint64, Level+1)
 	for i := 0; i < Level+1; i++ {
-		pol.Coeffs[i] = make([]uint64, N)
+		pol.Coeffs[i] = pol.Buff[i*N : (i+1)*N]
 	}
+
 	return
 }
 
@@ -35,19 +38,13 @@ func (pol *Poly) Level() int {
 
 // Zero sets all coefficients of the target polynomial to 0.
 func (pol *Poly) Zero() {
-	for _, coeffs := range pol.Coeffs {
-		ZeroVec(coeffs)
-	}
+	ZeroVec(pol.Buff)
 }
 
 // CopyNew creates an exact copy of the target polynomial.
 func (pol *Poly) CopyNew() (p1 *Poly) {
-	p1 = new(Poly)
-	p1.Coeffs = make([][]uint64, pol.Level()+1)
-	for i := range pol.Coeffs {
-		p1.Coeffs[i] = make([]uint64, len(pol.Coeffs[i]))
-		copy(p1.Coeffs[i], pol.Coeffs[i])
-	}
+	p1 = NewPoly(pol.N(), pol.Level())
+	copy(p1.Buff, pol.Buff)
 	p1.IsNTT = pol.IsNTT
 	p1.IsMForm = pol.IsMForm
 
@@ -58,31 +55,25 @@ func (pol *Poly) CopyNew() (p1 *Poly) {
 // Expects the degree of both polynomials to be identical.
 // Does not transfer the IsNTT and IsMForm flags.
 func CopyValues(p0, p1 *Poly) {
-	CopyValuesLvl(utils.MinInt(p0.Level(), p1.Level()), p0, p1)
+	copy(p1.Buff, p0.Buff)
+}
 
+// CopyValuesLvl copies the values of p0 on p1, up to level+1 moduli.
+func CopyValuesLvl(level int, p0, p1 *Poly) {
+	copy(p1.Buff[:p1.N()*(level+1)], p0.Buff)
 }
 
 // Copy copies the coefficients of p0 on p1 within the given Ring. It requires p1 to be at least as big p0.
 // Expects the degree of both polynomials to be identical.
 // Transfers the IsNTT and IsMForm flags.
 func Copy(p0, p1 *Poly) {
-	CopyValuesLvl(utils.MinInt(p0.Level(), p1.Level()), p0, p1)
+	CopyValues(p0, p1)
 	p1.IsNTT = p0.IsNTT
 	p1.IsMForm = p0.IsMForm
 }
 
-// CopyValuesLvl copies the coefficients of p0 on p1 within the given Ring for the moduli from 0 to level.
-// Expects the degree of both polynomials to be identical.
-// Does not transfer the IsNTT and IsMForm flags.
-func CopyValuesLvl(level int, p0, p1 *Poly) {
-	if p0 != p1 {
-		for i := 0; i < level+1; i++ {
-			copy(p1.Coeffs[i], p0.Coeffs[i])
-		}
-	}
-}
-
-// CopyLvl copies the coefficients of p0 on p1 within the given Ring for the moduli from 0 to level.
+// CopyLvl copies the coefficients of p0 on p1 within the given Ring.
+// Copies for up to level+1 moduli.
 // Expects the degree of both polynomials to be identical.
 // Transfers the IsNTT and IsMForm flags.
 func CopyLvl(level int, p0, p1 *Poly) {
@@ -97,10 +88,7 @@ func CopyLvl(level int, p0, p1 *Poly) {
 // Does not transfer the IsNTT and IsMForm flags.
 func (pol *Poly) CopyValues(p1 *Poly) {
 	if pol != p1 {
-		minLevel := utils.MinInt(pol.Level(), p1.Level())
-		for i := range p1.Coeffs[:minLevel+1] {
-			copy(pol.Coeffs[i], p1.Coeffs[i])
-		}
+		copy(pol.Buff, p1.Buff)
 	}
 }
 
@@ -122,38 +110,16 @@ func (pol *Poly) Equals(other *Poly) bool {
 	if pol == other {
 		return true
 	}
-	if pol != nil && other != nil && len(pol.Coeffs) == len(other.Coeffs) {
-		for i := range pol.Coeffs {
-			if len(other.Coeffs[i]) != len(pol.Coeffs[i]) {
+
+	if pol != nil && other != nil && len(pol.Buff) == len(other.Buff) {
+		for i := range pol.Buff {
+			if other.Buff[i] != pol.Buff[i] {
 				return false
-			}
-			for j := range pol.Coeffs[i] {
-				if other.Coeffs[i][j] != pol.Coeffs[i][j] {
-					return false
-				}
 			}
 		}
 		return true
 	}
 	return false
-}
-
-// SetCoefficients sets the coefficients of the polynomial directly from a CRT format (double slice).
-func (pol *Poly) SetCoefficients(coeffs [][]uint64) {
-	for i := range coeffs {
-		copy(pol.Coeffs[i], coeffs[i])
-	}
-}
-
-// GetCoefficients returns a new double slice that contains the coefficients of the polynomial.
-func (pol *Poly) GetCoefficients() (coeffs [][]uint64) {
-	coeffs = make([][]uint64, len(pol.Coeffs))
-	for i := range pol.Coeffs {
-		coeffs[i] = make([]uint64, len(pol.Coeffs[i]))
-		copy(coeffs[i], pol.Coeffs[i])
-	}
-
-	return
 }
 
 // GetDataLen64 returns the number of bytes a polynomial of N coefficients
@@ -206,7 +172,7 @@ func (pol *Poly) UnmarshalBinary(data []byte) (err error) {
 		return errors.New("invalid polynomial encoding")
 	}
 
-	if _, err = pol.DecodePoly64New(data); err != nil {
+	if _, err = pol.DecodePoly64(data); err != nil {
 		return err
 	}
 
@@ -237,23 +203,21 @@ func (pol *Poly) WriteTo64(data []byte) (int, error) {
 		data[6] = 1
 	}
 
-	return WriteCoeffsTo64(7, N, Level, pol.Coeffs, data)
+	return WriteCoeffsTo64(7, pol.Buff, data)
 }
 
 // WriteCoeffsTo64 converts a matrix of coefficients to a byte array, using 8 bytes per coefficient.
-func WriteCoeffsTo64(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 3
-	for i := 0; i < Level+1; i++ {
-		for j, k := 0, pointer; j < N; j, k = j+1, k+8 {
-			binary.BigEndian.PutUint64(data[k:], coeffs[i][j])
-		}
-		pointer += tmp
+func WriteCoeffsTo64(pointer int, coeffs []uint64, data []byte) (int, error) {
+	for i, j := 0, pointer; i < len(coeffs); i, j = i+1, j+8 {
+		binary.BigEndian.PutUint64(data[j:], coeffs[i])
 	}
 
-	return pointer, nil
+	return pointer + len(coeffs)*8, nil
 }
 
 // DecodePoly64 decodes a slice of bytes in the target polynomial returns the number of bytes decoded.
+// The method will first try to write on the buffer. If this step fails, either because the buffer isn't
+// allocated or of the wrong size, the method will allocate the correct buffer.
 // Assumes that each coefficient is encoded on 8 bytes.
 func (pol *Poly) DecodePoly64(data []byte) (pointer int, err error) {
 
@@ -270,36 +234,18 @@ func (pol *Poly) DecodePoly64(data []byte) (pointer int, err error) {
 
 	pointer = 7
 
-	if pointer, err = DecodeCoeffs64(pointer, N, Level, pol.Coeffs, data); err != nil {
+	if pol.Buff == nil || len(pol.Buff) != N*(Level+1) {
+		pol.Buff = make([]uint64, N*(Level+1))
+	}
+
+	if pointer, err = DecodeCoeffs64(pointer, pol.Buff, data); err != nil {
 		return pointer, err
 	}
 
-	return pointer, nil
-}
-
-// DecodePoly64New decodes a slice of bytes in the target polynomial returns the number of bytes decoded.
-// Assumes that each coefficient is encoded on 8 bytes.
-func (pol *Poly) DecodePoly64New(data []byte) (pointer int, err error) {
-
-	N := int(binary.BigEndian.Uint32(data))
-	Level := int(data[4])
-
-	if data[5] == 1 {
-		pol.IsNTT = true
-	}
-
-	if data[6] == 1 {
-		pol.IsMForm = true
-	}
-
-	pointer = 7
-
-	if pol.Coeffs == nil {
-		pol.Coeffs = make([][]uint64, Level+1)
-	}
-
-	if pointer, err = DecodeCoeffs64New(pointer, N, Level, pol.Coeffs, data); err != nil {
-		return pointer, err
+	// Reslice
+	pol.Coeffs = make([][]uint64, Level+1)
+	for i := 0; i < Level+1; i++ {
+		pol.Coeffs[i] = pol.Buff[i*N : (i+1)*N]
 	}
 
 	return pointer, nil
@@ -307,31 +253,12 @@ func (pol *Poly) DecodePoly64New(data []byte) (pointer int, err error) {
 
 // DecodeCoeffs64 converts a byte array to a matrix of coefficients.
 // Assumes that each coefficient is encoded on 8 bytes.
-func DecodeCoeffs64(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 3
-	for i := 0; i < Level+1; i++ {
-		for j, k := 0, pointer; j < N; j, k = j+1, k+8 {
-			coeffs[i][j] = binary.BigEndian.Uint64(data[k:])
-		}
-		pointer += tmp
+func DecodeCoeffs64(pointer int, coeffs []uint64, data []byte) (int, error) {
+	for i, j := 0, pointer; i < len(coeffs); i, j = i+1, j+8 {
+		coeffs[i] = binary.BigEndian.Uint64(data[j:])
 	}
 
-	return pointer, nil
-}
-
-// DecodeCoeffs64New converts a byte array to a matrix of coefficients.
-// Assumes taht each coefficient is encoded on 8 bytes.
-func DecodeCoeffs64New(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 3
-	for i := 0; i < Level+1; i++ {
-		coeffs[i] = make([]uint64, N)
-		for j, k := 0, pointer; j < N; j, k = j+1, k+8 {
-			coeffs[i][j] = binary.BigEndian.Uint64(data[k:])
-		}
-		pointer += tmp
-	}
-
-	return pointer, nil
+	return pointer + len(coeffs)*8, nil
 }
 
 // WriteTo32 writes the given poly to the data array.
@@ -356,23 +283,18 @@ func (pol *Poly) WriteTo32(data []byte) (int, error) {
 		data[6] = 1
 	}
 
-	cnt, err := WriteCoeffsTo32(7, N, Level, pol.Coeffs, data)
+	cnt, err := WriteCoeffsTo32(7, pol.Buff, data)
 
 	return cnt, err
 }
 
-// WriteCoeffsTo32 converts a matrix of coefficients to a byte array.
-// Encodes each coefficient on 4 bytes.
-func WriteCoeffsTo32(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 2
-	for i := 0; i < Level+1; i++ {
-		for j := 0; j < N; j++ {
-			binary.BigEndian.PutUint32(data[pointer+(j<<2):pointer+((j+1)<<2)], uint32(coeffs[i][j]))
-		}
-		pointer += tmp
+// WriteCoeffsTo32 converts a matrix of coefficients to a byte array, using 4 bytes per coefficient.
+func WriteCoeffsTo32(pointer int, coeffs []uint64, data []byte) (int, error) {
+	for i, j := 0, pointer; i < len(coeffs); i, j = i+1, j+4 {
+		binary.BigEndian.PutUint32(data[j:], uint32(coeffs[i]))
 	}
 
-	return pointer, nil
+	return pointer + len(coeffs)*4, nil
 }
 
 // GetPolyDataLen32 returns the number of bytes a polynomial of N coefficients
@@ -397,8 +319,9 @@ func (pol *Poly) GetDataLen32(WithMetadata bool) (cnt int) {
 }
 
 // DecodePoly32 decodes a slice of bytes in the target polynomial returns the number of bytes decoded.
-// Assumes that each coefficient is encoded on 4 bytes.
-// Writes on pre-allocated coefficients.
+// The method will first try to write on the buffer. If this step fails, either because the buffer isn't
+// allocated or of the wrong size, the method will allocate the correct buffer.
+// Assumes that each coefficient is encoded on 8 bytes.
 func (pol *Poly) DecodePoly32(data []byte) (pointer int, err error) {
 
 	N := int(binary.BigEndian.Uint32(data))
@@ -414,37 +337,18 @@ func (pol *Poly) DecodePoly32(data []byte) (pointer int, err error) {
 
 	pointer = 7
 
-	if pointer, err = DecodeCoeffs32(pointer, N, Level, pol.Coeffs, data); err != nil {
+	if pol.Buff == nil || len(pol.Buff) != N*(Level+1) {
+		pol.Buff = make([]uint64, N*(Level+1))
+	}
+
+	if pointer, err = DecodeCoeffs32(pointer, pol.Buff, data); err != nil {
 		return pointer, err
 	}
 
-	return pointer, nil
-}
+	pol.Coeffs = make([][]uint64, Level+1)
 
-// DecodePoly32New decodes a slice of bytes in the target polynomial returns the number of bytes decoded.
-// Assumes that each coefficient is encoded on 4 bytes.
-// Allocates the coefficients.
-func (pol *Poly) DecodePoly32New(data []byte) (pointer int, err error) {
-
-	N := int(binary.BigEndian.Uint32(data))
-	Level := int(data[4])
-
-	if data[5] == 1 {
-		pol.IsNTT = true
-	}
-
-	if data[6] == 1 {
-		pol.IsMForm = true
-	}
-
-	pointer = 7
-
-	if pol.Coeffs == nil {
-		pol.Coeffs = make([][]uint64, Level+1)
-	}
-
-	if pointer, err = DecodeCoeffs32New(pointer, N, Level, pol.Coeffs, data); err != nil {
-		return pointer, err
+	for i := 0; i < Level+1; i++ {
+		pol.Coeffs[i] = pol.Buff[i*N : (i+1)*N]
 	}
 
 	return pointer, nil
@@ -452,29 +356,10 @@ func (pol *Poly) DecodePoly32New(data []byte) (pointer int, err error) {
 
 // DecodeCoeffs32 converts a byte array to a matrix of coefficients.
 // Assumes that each coefficient is encoded on 4 bytes.
-func DecodeCoeffs32(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 2
-	for i := 0; i < Level+1; i++ {
-		for j, k := 0, pointer; j < N; j, k = j+1, k+4 {
-			coeffs[i][j] = uint64(binary.BigEndian.Uint32(data[k:]))
-		}
-		pointer += tmp
+func DecodeCoeffs32(pointer int, coeffs []uint64, data []byte) (int, error) {
+	for i, j := 0, pointer; i < len(coeffs); i, j = i+1, j+4 {
+		coeffs[i] = uint64(binary.BigEndian.Uint32(data[j:]))
 	}
 
-	return pointer, nil
-}
-
-// DecodeCoeffs32New converts a byte array to a matrix of coefficients.
-// Assumes that each coefficient is encoded on 4 bytes.
-func DecodeCoeffs32New(pointer, N, Level int, coeffs [][]uint64, data []byte) (int, error) {
-	tmp := N << 2
-	for i := 0; i < Level+1; i++ {
-		coeffs[i] = make([]uint64, N)
-		for j, k := 0, pointer; j < N; j, k = j+1, k+4 {
-			coeffs[i][j] = uint64(binary.BigEndian.Uint32(data[k:]))
-		}
-		pointer += tmp
-	}
-
-	return pointer, nil
+	return pointer + len(coeffs)*4, nil
 }