+++ rlwe/evaluator.go
@@ -1,9 +1,10 @@
 package rlwe
 
 import (
+	"math/bits"
+
 	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/rlwe/ringqp"
-	"math/bits"
 )
 
 // Evaluator is a struct that holds the necessary elements to execute general homomorphic