+++ rlwe/evaluator_automorphism.go
@@ -2,6 +2,7 @@ package rlwe
 
 import (
 	"fmt"
+
 	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/rlwe/ringqp"
 	"github.com/tuneinsight/lattigo/v3/utils"