+++ rlwe/lut/evaluator.go
@@ -1,11 +1,12 @@
 package lut
 
 import (
+	"math/big"
+
 	"github.com/tuneinsight/lattigo/v3/ring"
 	"github.com/tuneinsight/lattigo/v3/rlwe"
 	"github.com/tuneinsight/lattigo/v3/rlwe/rgsw"
 	"github.com/tuneinsight/lattigo/v3/rlwe/ringqp"
-	"math/big"
 )
 
 // Evaluator is a struct that stores necessary