+++ rlwe/lut/utils.go
@@ -1,8 +1,9 @@
 package lut
 
 import (
-	"github.com/tuneinsight/lattigo/v3/ring"
 	"math/big"
+
+	"github.com/tuneinsight/lattigo/v3/ring"
 )
 
 //MulBySmallMonomialMod2N multiplies pol by x^n, with 0 <= n < N